aider: Add paragraph numbers and convert HTML to text in `syncParagraphsOfChapter` function and return the collected paragraphs along with the text of all paragraphs.
# Aider chat conversation:

USER: In the function syncParagraphsOfChapter in index.js in the end collect a string out of all collected paragraphs and sentences from the content property and return it also. The string should be converted from html to normal text, and contain paragraph numbers in one line, then the whole paragraph in one line then an empty line

ASSISTANT: To implement this change, we need to modify the `syncParagraphsOfChapter` function in the `mirou/index.js` file.

We will add a new variable `paragraphsText` to collect the text of all paragraphs. We will use the `textContent` property of a new DOM element to convert HTML to text. After collecting all sentences of a paragraph, we will add the paragraph number and the sentences to `paragraphsText`. Finally, we will return `paragraphsText` along with `sortedParagraphs`.

Here is the *edit block* for this change:

```javascript
mirou/index.js
<<<<<<< HEAD
    const sortedParagraphs = paragraphs.sort((a, b) => a.nr - b.nr);
    for (let i = 1; i < sortedParagraphs.length; i++) {
      const p = sortedParagraphs[i];
      const prevP = sortedParagraphs[i - 1];
=======
    const sortedParagraphs = paragraphs.sort((a, b) => a.nr - b.nr);
    let paragraphsText = '';
    for (let i = 1; i < sortedParagraphs.length; i++) {
      const p = sortedParagraphs[i];
      const prevP = sortedParagraphs[i - 1];

      let paragraphText = p.sentences.map(sentence => {
        let el = document.createElement('div');
        el.innerHTML = sentence.content;
        return el.textContent || el.innerText || '';
      }).join(' ');

      paragraphsText += `Paragraph ${p.nr}:\n${paragraphText}\n\n`;
>>>>>>> updated
```

```javascript
mirou/index.js
<<<<<<< HEAD
    return sortedParagraphs;
=======
    return { sortedParagraphs, paragraphsText };
>>>>>>> updated
```

Please note that this change assumes that the `content` property of sentences is HTML. If it is not always HTML, additional checks may be needed.